Fx-fix

+++ src/main/java/sausegeShop/Fx/MainMenu.java
@@ -59,7 +59,7 @@ public class MainMenu {
         });
     }
 
-    private void findProduct(){
+    private void findProduct() {
         find.setOnAction(actionEvent -> {
             String product = textToFind.getText();
             if (product.contains("?")) {
@@ -67,10 +67,11 @@ public class MainMenu {
                 actualSearch = actualSearch + ".+.";
                 for (int i = 0; i < productController.size(); i++) {
                     if (Pattern.matches(actualSearch, productController.getProduct(i).getName())) {
-                        Button button = new Button(productController.getProduct(i).getName()+ "    "+ productController.getProduct(i).getPrice());
+                        Button button = new Button(productController.getProduct(i).getName() + "    " + productController.getProduct(i).getPrice());
                         button.setMinSize(375, 50);
                         int finalI = i;
                         button.setOnAction(actionEvent1 -> {
+                            showSome.getChildren().clear();
                             showOneProduct(productController.getProduct(finalI));
                         });
                         showSome.getChildren().add(button);
@@ -79,10 +80,11 @@ public class MainMenu {
             } else {
                 for (int i = 0; i < productController.size(); i++) {
                     if (productController.getProduct(i).getName().toLowerCase().contains(product.toLowerCase())) {
-                        Button button = new Button(productController.getProduct(i).getName()+ "    "+ productController.getProduct(i).getPrice());
+                        Button button = new Button(productController.getProduct(i).getName() + "    " + productController.getProduct(i).getPrice());
                         int finalI = i;
                         button.setMinSize(375, 50);
                         button.setOnAction(actionEvent1 -> {
+                            showSome.getChildren().clear();
                             showOneProduct(productController.getProduct(finalI));
                         });
                         showSome.getChildren().add(button);
@@ -95,36 +97,42 @@ public class MainMenu {
     private void showBasket() {
         showSome.getChildren().clear();
         cart.setOnAction(e -> {
-            for (int i = 0; i < basketController.getBasket().size(); i++) {
-                Label name = new Label("Название: " + basketController.getBasket().getProducts(i).getName());
-                Label price = new Label("Цена: " + basketController.getBasket().getProducts(i).getPrice());
-                Label description = new Label("Описание: " + basketController.getBasket().getProducts(i).getDescription());
-                Label composition = new Label("Состав: " + basketController.getBasket().getProducts(i).getComposition());
-                Label count = new Label("Количество: " + basketController.getBasket().getCount(i));
-                Label rat = new Label("Рейтинг: " + basketController.getBasket().getRat(i));
-                Button remove = new Button("Удалить");
-                int finalI = i;
-                //Здесь происходит удаление товара, неправильно убирает надписи
-                remove.setOnAction(actionEvent -> {
-                    basketController.getBasket().delete(finalI);
-                    if (finalI == 0)
-                        showSome.getChildren().remove(0, 7);
-                    else
-                        showSome.getChildren().remove(7 * finalI, 7 * finalI + 7);
+            if (basketController.getBasket().size() == 0) {
+                Label label = new Label("В коризине пусто");
+                showSome.getChildren().add(label);
+            } else {
+                for (int i = 0; i < basketController.getBasket().size(); i++) {
+                    Label name = new Label("Название: " + basketController.getBasket().getProducts(i).getName());
+                    Label price = new Label("Цена: " + basketController.getBasket().getProducts(i).getPrice());
+                    Label description = new Label("Описание: " + basketController.getBasket().getProducts(i).getDescription());
+                    Label composition = new Label("Состав: " + basketController.getBasket().getProducts(i).getComposition());
+                    Label count = new Label("Количество: " + basketController.getBasket().getCount(i));
+                    Label rat = new Label("Рейтинг: " + basketController.getBasket().getRat(i));
+                    Button remove = new Button("Удалить");
+                    int finalI = i;
+                    //Здесь происходит удаление товара, неправильно убирает надписи
+                    remove.setOnAction(actionEvent -> {
+                        basketController.getBasket().delete(finalI);
+                        if (finalI == 0)
+                            showSome.getChildren().remove(0, 7);
+                        else
+                            showSome.getChildren().remove(7 * finalI, 7 * finalI + 7);
+                    });
+                    showSome.getChildren().addAll(name, price, description, composition, count, rat, remove);
+                }
+                Button buyAll = new Button("Купить все");
+                buyAll.setOnAction(actionEvent -> {
+                    Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                    alert.setTitle("Спасибо");
+                    alert.setContentText("Спаибо за покупку");
+                    alert.showAndWait();
+                    basketController.getBasket().deleteAll();
                 });
-                showSome.getChildren().addAll(name, price, description, composition, count, rat, remove);
+                showSome.getChildren().add(buyAll);
             }
-            Button buyAll = new Button("Купить все");
-            buyAll.setOnAction(actionEvent -> {
-                Alert alert = new Alert(Alert.AlertType.INFORMATION);
-                alert.setTitle("Спасибо");
-                alert.setContentText("Спаибо за покупку");
-                alert.showAndWait();
-                basketController.getBasket().deleteAll();
-            });
-            showSome.getChildren().add(buyAll);
         });
     }
+
 
     private void goToAdminMenu() {
         adminMenu.setOnAction(e -> {
@@ -153,6 +161,7 @@ public class MainMenu {
                 Button button = new Button();
                 int finalI = i;
                 button.setOnAction(event -> {
+                    showSome.getChildren().clear();
                     showProducts(categoryController.getCategory(finalI));
                 });
                 button.setId(String.valueOf(i));
@@ -169,6 +178,7 @@ public class MainMenu {
             Button button = new Button();
             int finalI = i;
             button.setOnAction(actionEvent -> {
+                showSome.getChildren().clear();
                 showOneProduct(category.getProduct(finalI));
             });
             button.setId(String.valueOf(i));
@@ -207,6 +217,7 @@ public class MainMenu {
             }
         });
         back.setOnAction(actionEvent -> {
+            showSome.getChildren().clear();
             showProducts(product.getCategory());
         });
     }
